Adds support for spike detection and fixing
We adds functions for detecting noise artifact count spikes, and for fixing them with the mean of their neighbour bin. At present, the function for finding noise spikes does not ch
eck the first and last bin: it's not crucial but we should be mindful of that and consider correcting for it later. These new functions come with tests.