Improves moving average logic
Now we look first for a valid data range to run centered moving average on. Removed asymmetric edge logic because it would result in significant biases when data starts or ends with a trend, which is common because the instrument is powered off at polar regions. To deal with background and count alignment without introducing and checking on NaNs, the trigger logic requires the range specification with other inputs.